# web/Openapi/types.gen.go
// Package Openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.8.2 DO NOT EDIT.
package Openapi

// Todo defines model for Todo.
type Todo struct {
	Content   string  `json:"content"`
	CreatedAt string  `json:"createdAt"`
	Done      bool    `json:"done"`
	Id        int     `json:"id"`
	Note      *string `json:"note"`
	UpdatedAt string  `json:"updatedAt"`
	UserId    int     `json:"userId"`
}

// User defines model for User.
type User struct {
	CreatedAt string  `json:"createdAt"`
	Email     string  `json:"email"`
	Id        int     `json:"id"`
	Name      string  `json:"name"`
	Note      *string `json:"note"`
	Password  string  `json:"password"`
	UpdatedAt string  `json:"updatedAt"`
}

// PostUserJSONBody defines parameters for PostUser.
type PostUserJSONBody struct {
	Email    string  `json:"email"`
	Name     string  `json:"name"`
	Note     *string `json:"note"`
	Password string  `json:"password"`
}

// PatchUsersUserIdJSONBody defines parameters for PatchUsersUserId.
type PatchUsersUserIdJSONBody struct {
	Email    *string `json:"email,omitempty"`
	Id       int     `json:"id"`
	Name     *string `json:"name,omitempty"`
	Password *string `json:"password,omitempty"`
}

// PostUsersUserIdTodoJSONBody defines parameters for PostUsersUserIdTodo.
type PostUsersUserIdTodoJSONBody struct {
	Content string  `json:"content"`
	Note    *string `json:"note"`
	UserId  int     `json:"userId"`
}

// PatchUsersUserIdTodoTodoIdJSONBody defines parameters for PatchUsersUserIdTodoTodoId.
type PatchUsersUserIdTodoTodoIdJSONBody struct {
	Content *string `json:"content,omitempty"`
	Done    *bool   `json:"done,omitempty"`
	Id      int     `json:"id"`
	Note    *string `json:"note,omitempty"`
}

// PostUserJSONRequestBody defines body for PostUser for application/json ContentType.
type PostUserJSONRequestBody PostUserJSONBody

// PatchUsersUserIdJSONRequestBody defines body for PatchUsersUserId for application/json ContentType.
type PatchUsersUserIdJSONRequestBody PatchUsersUserIdJSONBody

// PostUsersUserIdTodoJSONRequestBody defines body for PostUsersUserIdTodo for application/json ContentType.
type PostUsersUserIdTodoJSONRequestBody PostUsersUserIdTodoJSONBody

// PatchUsersUserIdTodoTodoIdJSONRequestBody defines body for PatchUsersUserIdTodoTodoId for application/json ContentType.
type PatchUsersUserIdTodoTodoIdJSONRequestBody PatchUsersUserIdTodoTodoIdJSONBody